requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
langchain==0.1.0
chromadb==0.4.22
sentence-transformers==2.6.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
tqdm==4.66.1 